+++ src/main/java/org/example/diplomwork/entities/Position.java
@@ -1,26 +1,26 @@
 package org.example.diplomwork.entities;
 
 public class Position {
-    private Integer x;
-    private Integer y;
+    private Double x;
+    private Double y;
 
     public Position() {
-        this.x = 0;
-        this.y = 0;
+        this.x = 0.;
+        this.y = 0.;
     }
 
-    public Position(Integer x, Integer y) {
+    public Position(Double x, Double y) {
         this.x = x;
         this.y = y;
     }
 
-    public Integer getX() { return x; }
+    public Double getX() { return x; }
 
-    public Integer getY() { return y; }
+    public Double getY() { return y; }
 
-    public void setX(Integer x) { this.x = x; }
+    public void setX(Double x) { this.x = x; }
 
-    public void setY(Integer y) { this.y = y; }
+    public void setY(Double y) { this.y = y; }
 
     @Override
     public String toString() {
